WebSockets/tests/submissions/Opera/interfaces/WebSocket/send/002.html

<!doctype html>
<title>WebSockets: replacing send</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js></script>
<div id=log></div>
<script>
test(function() {
  var ws = new WebSocket(SCHEME_AND_DOMAIN+'/');
  ws.send = 5;
  assert_equals(ws.send, 5);
});
</script>
